Move W3CTraceParent to internal-api, rather than have it both as an injected helper and an isolated tracer class. This should avoid an obscure error when generating native-images.
Also remove inject requests for datadog.trace.bootstrap.instrumentation.dbm.SharedDBCommenter as it is implicitly available on the boot-class-path.
