+++ dd-java-agent/instrumentation/jdbc/src/main/java/datadog/trace/instrumentation/jdbc/AbstractPreparedStatementInstrumentation.java
@@ -40,7 +40,7 @@ public abstract class AbstractPreparedStatementInstrumentation extends Instrumen
   @Override
   public String[] helperClassNames() {
     return new String[] {
-      "datadog.trace.core.propagation.W3CTraceParent", packageName + ".JDBCDecorator",
+      packageName + ".JDBCDecorator",
     };
   }
