+++ dd-java-agent/instrumentation/jdbc/src/main/java/datadog/trace/instrumentation/jdbc/DBMCompatibleConnectionInstrumentation.java
@@ -79,10 +79,7 @@ public class DBMCompatibleConnectionInstrumentation extends AbstractConnectionIn
   @Override
   public String[] helperClassNames() {
     return new String[] {
-      "datadog.trace.core.propagation.W3CTraceParent",
-      packageName + ".JDBCDecorator",
-      packageName + ".SQLCommenter",
-      "datadog.trace.bootstrap.instrumentation.dbm.SharedDBCommenter",
+      packageName + ".JDBCDecorator", packageName + ".SQLCommenter",
     };
   }
