+++ dd-java-agent/instrumentation/jdbc/src/main/java/datadog/trace/instrumentation/jdbc/DriverInstrumentation.java
@@ -49,7 +49,7 @@ public final class DriverInstrumentation extends InstrumenterModule.Tracing
   @Override
   public String[] helperClassNames() {
     return new String[] {
-      "datadog.trace.core.propagation.W3CTraceParent", packageName + ".JDBCDecorator",
+      packageName + ".JDBCDecorator",
     };
   }
