+++ dd-java-agent/instrumentation/mongo/mongo-driver/mongo-driver-3/mongo-driver-3.6/src/main/java/datadog/trace/instrumentation/mongo/DefaultServerConnection36Instrumentation.java
@@ -32,11 +32,9 @@ public class DefaultServerConnection36Instrumentation extends InstrumenterModule
   @Override
   public String[] helperClassNames() {
     return new String[] {
-      "datadog.trace.core.propagation.W3CTraceParent",
       packageName + ".BsonScrubber",
       packageName + ".MongoCommentInjector",
       packageName + ".MongoDecorator",
-      "datadog.trace.bootstrap.instrumentation.dbm.SharedDBCommenter",
     };
   }
