+++ dd-java-agent/instrumentation/mongo/mongo-driver/mongo-driver-3/mongo-driver-3.8/src/main/java/datadog/trace/instrumentation/mongo/DefaultServerConnection38Instrumentation.java
@@ -32,8 +32,6 @@ public class DefaultServerConnection38Instrumentation extends InstrumenterModule
   @Override
   public String[] helperClassNames() {
     return new String[] {
-      "datadog.trace.bootstrap.instrumentation.dbm.SharedDBCommenter",
-      "datadog.trace.core.propagation.W3CTraceParent",
       packageName + ".MongoDecorator",
       packageName + ".MongoCommentInjector",
       packageName + ".BsonScrubber",
